Look for templates in Application Support before searching within bundled templates
